Deep planning, deep exploration, and impossible side effects learning
Deep planning works. I already forgot how many time steps are necessary, but it is something on the order of 1e5 I think?

Deep exploration works but is very slow.
I needed something like 2e9 time steps.
Maybe it converges before that, but it is difficult to read the meassive data file.
I've stored it temporarily as deep_exploration locally.

I don't get convergence when there are intracellular states for which side effects labels are absent.
Could be because convergence takes time and 4e6 time steps just isnt enough for 27 states and 27 actions.
It doesnt reach the states where it thinks the most value is, but that is to be expected since the transition fn is not yet learnt.
It does however linger in states where r_distances is big, so that suggests exploration is still happening.
Maybe I should make cells2states3actions3?