Found one sufficient explanation for failed verification
The output was not saved in the right place. Now that it is, I see that the intervals are incorrect. In particular, I have found that the p_estimates are sometimes too large. Now I have to find out where they grow so large. They should never be larger than 1.