Debugging of comparison algorithms
The only comparison algorithm that seemed to be wrong was AlwaysSafe/PSO. The error was in confusing count and cell. That is fixed now, and results seem reasonable.

Apart from that results seem like one would expect them to be. Aup is suboptimal in terms of rewards and does not even do that well wrt side effects. Naton-like seems to work surprisingly well.
- [ ] make the environments more exploratorily challenging.